Create base heatmap data
